fix: fk defined joins when extended
